chore: migrate connector tests to bun test with real sqlite db; add deps injection, unregisterConnector, setupTestDb helper
